requirements.txt
streamlit>=1.33.0
python-pptx>=0.6.23
Pillow>=10.3.0
ImageHash>=4.3.1
PyMuPDF>=1.24.0
